fix errs; use instantaneous scores instead of ema scores to calc prev topic quantile score
